requirements.txt
torch
torchvision
flask
pyTelegramBotAPI
pillow